Update README.md
This is the user’s guide of the code developed for [1]. The idea of this work is fine-tuning and personalizing compression ratios of hearing aid compression.
For this reason, in order to decrease the number of user feedbacks and thus enable a practical implementation of the personalized compression, first a reward function is considered to model hearing preferences of a user in an asynchronous manner. This is achieved by carrying out comparison between instances of two different compressed audios. Then, an agent is trained to maximize reward. Following figure shows a block diagram of this approach:

 


Figure 1:  Block diagram of proposed approach.

Below shows how data is passes through different blocks of the personalization framework both in training and testing modes. Please refer to [1] for more details.


Figure 2. Developed personalized compression DRL framework for (a) training mode and (b) operation mode.